updating to reset_autocompleter_providers
